Auth Service/importers.py
import pymongo
import os
import xml
import json
import hashlib
import jwt

import xml.etree.ElementTree as ET


import logging
logging.basicConfig(level=logging.INFO)


from datetime import datetime,timedelta
from spyne import Application, rpc, ServiceBase,Integer, Unicode,Date
from spyne.model.complex import XmlData
from spyne import Iterable
from spyne.protocol.soap import Soap11
from spyne.server.wsgi import WsgiApplication
