Auth Service/requirements.txt
annotated-types==0.7.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
gunicorn==23.0.0
h11==0.14.0
lxml==5.3.1
MarkupSafe==3.0.2
packaging==24.2
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyJWT==2.10.1
pymongo==4.11.3
python-dotenv==1.1.0
pytz==2025.2
spyne==2.14.0
typing-inspection==0.4.0
typing_extensions==4.13.1
uvicorn==0.34.0
Werkzeug==3.1.3
